+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 0.10.0
 
+#### 💥 Breaking
+
+- Removed `install_global`, use `go install` instead.
+- Removed `uninstall_global`.
+
 #### 🚀 Updates
 
 - Updated to support proto v0.31 release.